Separate out email logics
