+++ MonitoringService/Program.cs
@@ -1,8 +1,8 @@
 using Microsoft.EntityFrameworkCore;
 using MonitoringService;
 using MonitoringService.Domain.Models;
-using MonitoringService.Emails;
 using MonitoringService.Persistence;
+using MonitoringService.Services;
 
 IHost host = Host.CreateDefaultBuilder(args)
     .ConfigureServices((context, services) =>
@@ -12,8 +12,9 @@ IHost host = Host.CreateDefaultBuilder(args)
             options.UseSqlServer(connectionString));
 
         services.Configure<ConfigurableSettings>(context.Configuration.GetSection("Options"));
-        services.Configure<EmailSettings>(context.Configuration.GetSection("EmailSettings"));
+        services.Configure<EmailProperties.EmailSettings>(context.Configuration.GetSection("EmailSettings"));
         services.AddSingleton<EmailService>();
+        services.AddTransient<FileDirectorySetup>();
         services.AddHostedService<Worker>();
     })
     .Build();
